backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
